app.py
import streamlit as st
import requests
from PIL import Image
import io
import time

# Page Configuration
st.set_page_config(
    page_title="Smart Plate Scanner",
    page_icon="🚗",
    layout="wide",
    initial_sidebar_state="expanded"
)

# Custom CSS
st.markdown("""
    <style>
    /* Main content styling */
    .main .block-container {
        max-width: 1400px;
        padding: 2rem 3rem;
        background-color: #f8f9fa;
    }
    
    /* Headers */
    h1, h2, h3 {
        color: #2c3e50;
        font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
        font-weight: 700;
        margin-bottom: 1rem;
    }
    
    /* Info box styling */
    .info-box {
        background-color: #ffffff;
        border: 1px solid #dfe6e9;
        border-radius: 12px;
        padding: 1.5rem;
        margin: 1rem 0;
        box-shadow: 0 4px 6px rgba(0,0,0,0.05);
    }
    
    /* Button styling */
    .stButton>button {
        background-color: #3498db;
        color: white;
        border: none;
        padding: 0.75rem 1.5rem;
        border-radius: 8px;
        font-size: 1rem;
        font-weight: 600;
        transition: all 0.3s ease;
        width: 100%;
        margin: 1rem 0;
    }
    
    .stButton>button:hover {
        background-color: #2980b9;
        transform: translateY(-2px);
        box-shadow: 0 4px 8px rgba(0,0,0,0.1);
    }
    
    /* File uploader styling */
    .stFileUploader {
        border: 2px dashed #bdc3c7;
        border-radius: 12px;
        padding: 2rem;
        background-color: #ffffff;
        margin: 1rem 0;
    }
    
    /* Result card styling */
    .result-card {
        background-color: #ffffff;
        border: 1px solid #dfe6e9;
        border-radius: 12px;
        padding: 1.5rem;
        margin: 1rem 0;
        box-shadow: 0 4px 6px rgba(0,0,0,0.05);
    }
    
    /* Image container */
    .image-container {
        background: white;
        padding: 1rem;
        border-radius: 12px;
        box-shadow: 0 4px 6px rgba(0,0,0,0.05);
        margin-bottom: 1.5rem;
        border: 1px solid #dfe6e9;
        text-align: center;
    }
    
    /* Footer styling */
    .footer {
        text-align: center;
        padding: 1.5rem;
        color: #7f8c8d;
        font-size: 0.9rem;
        margin-top: 2rem;
        border-top: 1px solid #dfe6e9;
    }
    
    /* Responsive columns */
    @media (max-width: 768px) {
        .main .block-container {
            padding: 1rem;
        }
    }
    </style>
""", unsafe_allow_html=True)

# Sidebar
with st.sidebar:
    st.markdown("""
        <div style='text-align: center; margin-bottom: 2rem;'>
            <h2>🚗 Plate Scanner</h2>
            <p style='color: #7f8c8d;'>Upload vehicle image for license plate recognition</p>
        </div>
    """, unsafe_allow_html=True)
  
    st.markdown("### How to Use")
    st.markdown("""
        <div style='color: #7f8c8d;'>
            <ol>
                <li>Upload vehicle image or take photo</li>
                <li>Click 'Analyze Image'</li>
                <li>View detected license plate</li>
            </ol>
        </div>
    """, unsafe_allow_html=True)

# Main content
st.title("Smart License Plate Recognition")
st.markdown("""
    <p style='color: #7f8c8d; font-size: 1.1rem;'>
        This application uses AI to detect and recognize license plates from vehicle images.
        Upload an image below to get started.
    </p>
""", unsafe_allow_html=True)

# Main content layout
tab1, tab2 = st.tabs(["📷 Image Analysis", "ℹ️ System Info"])

with tab1:
    col1, col2 = st.columns([2, 1])
    
    with col1:
        st.markdown("### Image Input")
        with st.container():
            source = st.radio("Select input method:", ("Upload Image", "Take Photo"), horizontal=True)
            
            if source == "Upload Image":
                uploaded_file = st.file_uploader(
                    "Choose an image file (PNG, JPG, JPEG)",
                    type=['png', 'jpg', 'jpeg'],
                    label_visibility="collapsed"
                )
            else:
                uploaded_file = st.camera_input("Take a photo", label_visibility="collapsed")
            
            if uploaded_file is not None:
                # Display the uploaded image
                image = Image.open(uploaded_file)
                st.markdown("#### Image Preview")
                st.image(image, caption='Uploaded Image', use_container_width=True)

    with col2:
        if uploaded_file:
            try:
                if st.button("Analyze Image", type="primary"):
                    with st.spinner("Processing image..."):
                        # Convert image to bytes
                        img_bytes = io.BytesIO()
                        image.save(img_bytes, format='PNG')
                        
                        # API Request
                        response = requests.post(
                            "http://localhost:8000/predict",
                            files={"file": ("scan.png", img_bytes.getvalue(), "image/png")}
                        )
                        
                        if response.status_code == 200:
                            result = response.json().get('number', '')
                            confidence = response.json().get('confidence', 0)
                            
                            st.markdown("### Analysis Results")
                            with st.container():
                                st.markdown(f"""
                                    <div class='result-card'>
                                        <h4 style='color: #000000;'>Detected License Plate</h4>
                                        <p style='font-size: 2rem; color: #000000; font-weight: bold;'>{result}</p>
                                        # <p style='color: #7f8c8d;'>Confidence: {confidence:.1f}%</p>
                                    </div>
                                """, unsafe_allow_html=True)
                        else:
                            st.error("Analysis failed. Please try again.")
            except Exception as e:
                st.error(f"Error processing image: {str(e)}")

with tab2:
    st.markdown("""
        ### About This System  
        This license plate recognition system uses advanced computer vision and 
        machine learning techniques to identify vehicle license plates.
        
        #### Technical Features  
        - Image preprocessing pipeline  
        - Optical character recognition (OCR)  
        - Confidence scoring system  
        - REST API integration  
        
        #### Performance Metrics  
        - 95% accuracy on standard test set  
        - 500ms average processing time  
        - Supports multiple plate formats  
        
        #### Applications  
        - Parking management systems  
        - Traffic monitoring  
        - Security systems  
        - Toll collection automation  
    """, unsafe_allow_html=True)

# Footer
st.markdown("""
    <div class="footer">
        <p>Smart Plate Scanner v1.0 | Secure AI Processing</p>
    </div>
""", unsafe_allow_html=True)