requirements.txt
streamlit==1.32.0
requests==2.31.0
Pillow==10.2.0
fastapi==0.109.2
python-multipart==0.0.9
uvicorn==0.27.1 